移除request与response类
